requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
APScheduler==3.10.4
aiohttp==3.12.15
aiohttp-retry==2.9.1
